perf: consolidate benchmark into single Python script, add dated findings
Replace hyperfine + bash scripts with a single bench.py that handles both
sequential and concurrent modes. Add FINDINGS-2026-06-20-mbp.md with full
results from MacBook Pro (M1 Pro, 32GB) including hardware/runtime specs.

Key result: fogwall 3.2x faster on concurrent push (3.5 vs 1.1 ops/s at
10 parallel / 50 ops). Clone and fetch within noise of direct Gitea for
both proxies.
