+++ perf/README.md
@@ -7,68 +7,88 @@ not database or network latency.
 
 ## Prerequisites
 
-- Docker (with Compose v2)
-- [hyperfine](https://github.com/sharkdp/hyperfine) — `brew install hyperfine`
+- Docker (for Gitea only — pre-built image, no build needed)
+- Python 3 (standard library only, no pip packages)
+- Java 25+ (for fogwall)
+- Node.js 18+ (for finos/git-proxy)
 - git
 
 ## Quick start
 
 ```bash
-# 1. Build and start all services (Gitea + fogwall + finos/git-proxy)
-docker compose -f perf/docker-compose.yml up -d --build
+# 1. Start services locally
+bash perf/start-local.sh
 
 # 2. One-time setup: create test org, repo, user, seed data
+bash perf/setup.sh --local
+
+# 3. Run sequential benchmarks
+python3 perf/bench.py fogwall
+python3 perf/bench.py git-proxy
+
+# 4. Run concurrent benchmarks
+CONCURRENCY=20 TOTAL_OPS=100 python3 perf/bench.py fogwall --concurrent
+CONCURRENCY=20 TOTAL_OPS=100 python3 perf/bench.py git-proxy --concurrent
+
+# 5. Stop services
+bash perf/stop-local.sh
+```
+
+## Docker Compose (fully isolated)
+
+```bash
+docker compose -f perf/docker-compose.yml up -d --build
 bash perf/setup.sh
+python3 perf/bench.py fogwall
+docker compose -f perf/docker-compose.yml down -v
+```
 
-# 3. Run benchmarks
-bash perf/bench.sh
+## Tuning
 
-# 4. View results
-cat perf/results/clone.md
-cat perf/results/fetch.md
-cat perf/results/push.md
+```bash
+# Sequential: more runs, more warmup
+RUNS=20 WARMUP=5 python3 perf/bench.py fogwall
+
+# Concurrent: scale up
+CONCURRENCY=50 TOTAL_OPS=200 python3 perf/bench.py fogwall --concurrent
 ```
 
 ## What's tested
 
-| Scenario | fogwall | finos/git-proxy |
-|----------|---------|-----------------|
-| `clone`  | transparent proxy (Jetty ProxyServlet) | transparent proxy (express-http-proxy) |
-| `fetch`  | transparent proxy | transparent proxy |
-| `push`   | transparent proxy (servlet filter chain inspects pack, then forwards) | transparent proxy (clones upstream into temp dir for inspection, deletes clone, then forwards original request) |
+| Scenario | Description |
+|----------|-------------|
+| Clone | Full `git clone` through proxy vs direct Gitea |
+| Fetch | `git fetch` on pre-cloned repo |
+| Push | Commit + push through proxy (transparent mode) |
 
-Each scenario also includes a **direct Gitea** baseline (no proxy) for clone
-and fetch to show raw overhead.
+### Push architecture difference
 
-A **rapid fetch** scenario (50 runs, no warmup) tests cache behavior — how
-efficiently each proxy serves repeated fetches.
-
-### Push inspection strategies
-
-Both proxies transparently forward the push to upstream, but inspect the pack
+Both proxies transparently forward pushes to upstream, but inspect the pack
 data differently:
 
-- **fogwall** parses the pack inline in the servlet filter chain (no local
-  clone needed for transparent mode). The filter chain reads the pack stream,
-  runs validation, and forwards to upstream in a single pass.
+- **fogwall** maintains a persistent cached clone per repo
+  (`LocalRepositoryCache`) that is reused across requests. On each push, it
+  does an incremental fetch (with a 5s cooldown to avoid redundant fetches),
+  inspects commits from the cache, and forwards the original request to
+  upstream.
 
-- **finos/git-proxy** clones the upstream repo into a temp directory, runs
-  `git receive-pack` locally to unpack and inspect, deletes the clone, then
-  forwards the original HTTP request to upstream. The extra clone+receive-pack
-  adds overhead that should show up in the push benchmark.
+- **finos/git-proxy** does a fresh `git clone` of the upstream repo into a
+  temp directory on every push, runs `git receive-pack` into it for
+  inspection, then deletes the clone and forwards the original request. Every
+  push pays the full clone cost.
 
 fogwall also supports a separate **store-and-forward mode** (not benchmarked
 here) where pushes are received into a persistent local bare repo and forwarded
 on behalf of the user — enabling deferred forwarding and approval gates.
 
-## Tuning
+## Results
 
-```bash
-RUNS=20 WARMUP=3 bash perf/bench.sh
-```
+See [results/FINDINGS.md](results/FINDINGS.md) for detailed benchmark results.
 
 ## Teardown
 
 ```bash
+bash perf/stop-local.sh
+# or for Docker Compose:
 docker compose -f perf/docker-compose.yml down -v
 ```
